moves parseOptions logic to pkg/launcher, updates flare to use func and dump file to log
